Adds -ca-name argument for setting root CA subj common name.
This commit adds a `-ca-name` argument to the minica command. If
specified the value is used to override the subject common name (CN) of
the root CA certificate. If not present, the existing default of "minica
root ca %s" is used, where %s is the first three hex bytes of the cert
serial.